requirements.txt
# Dependências principais
python-dotenv==1.0.0
twilio==8.10.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
sqlite3-utils==0.1
openai==1.3.0
python-multipart==0.0.6

# Dependências para desenvolvimento
pytest==7.4.3
black==23.11.0
flake8==6.1.0